+++ assets/js/core/constants.js
@@ -1,0 +1,13 @@
+export const TICKET_FIELD_LABELS = {
+  id:'ID',
+  createdAt:'Data de criação',
+  solicitante:'Aberto por',
+  telefone:'Contato',
+  meetPoint:'Ponto de encontro',
+  dupla:'Dupla',
+  concl:'% de conclusão',
+  dueDate:'Prazo (data final)',
+  resumo:'Resumo',
+  descricao:'Descrição'
+};
+export const TICKET_FIELDS = Object.keys(TICKET_FIELD_LABELS);
