+++ assets/js/tpl/index.js
@@ -1,0 +1,3 @@
+import { login } from './login.js';
+import { dashboard } from './dashboard.js';
+export const tpl = { login, dashboard };
