+++ assets/js/ui/admin/create-project.js
@@ -1,0 +1,3 @@
+export function showCreateProject(){
+  // placeholder create project form
+}
