+++ assets/js/ui/admin/create-ticket.js
@@ -1,0 +1,3 @@
+export function showCreateTicket(){
+  // placeholder create ticket form
+}
