+++ assets/js/ui/charts.js
@@ -1,0 +1,3 @@
+export function renderSLAChart(){
+  // placeholder charts render
+}
