+++ assets/js/ui/layout.js
@@ -1,0 +1,3 @@
+export function initLayout(){
+  // placeholder for greeting and clock setup
+}
