+++ assets/js/ui/projects/archives.js
@@ -1,0 +1,6 @@
+export function renderArchivedProjects(){
+  // placeholder for archived projects
+}
+export function renderFinishedProjects(){
+  // placeholder for finished projects
+}
