+++ assets/js/ui/projects/carousel.js
@@ -1,0 +1,3 @@
+export function renderProjects(){
+  // placeholder for project carousel
+}
