+++ assets/js/ui/projects/detail.js
@@ -1,0 +1,3 @@
+export function openProjectDetailInline(){
+  // placeholder for project detail
+}
