+++ assets/js/ui/projects/notes.js
@@ -1,0 +1,3 @@
+export function renderProjectNotes(){
+  // placeholder for project notes
+}
