+++ assets/js/ui/projects/observacoes.js
@@ -1,0 +1,3 @@
+export function renderProjectObservacoes(){
+  // placeholder for project observacoes
+}
