+++ assets/js/ui/router.js
@@ -1,0 +1,9 @@
+export function setActiveTab(id){
+  // placeholder router
+}
+export function hideAllSections(){
+  // placeholder
+}
+export function setPageState(){
+  // placeholder
+}
