+++ assets/js/ui/tickets/archives.js
@@ -1,0 +1,6 @@
+export function renderArchivedTickets(){
+  // placeholder for archived tickets
+}
+export function renderFinishedTickets(){
+  // placeholder for finished tickets
+}
