+++ assets/js/ui/tickets/detail.js
@@ -1,0 +1,3 @@
+export function openTicketDetail(){
+  // placeholder for ticket detail
+}
