+++ assets/js/ui/tickets/list.js
@@ -1,0 +1,3 @@
+export function renderTickets(){
+  // placeholder tickets list
+}
