+++ assets/js/ui/tickets/notes.js
@@ -1,0 +1,3 @@
+export function renderTicketNotes(){
+  // placeholder for ticket notes
+}
