+++ assets/js/ui/tickets/observacoes.js
@@ -1,0 +1,3 @@
+export function renderObservacoes(){
+  // placeholder for ticket observacoes
+}
