+++ assets/js/ui/tv-mode.js
@@ -1,0 +1,3 @@
+export function toggleTVMode(){
+  // placeholder fullscreen toggle
+}
