+++ index.html
@@ -48,9 +48,6 @@
   <div class="project-detail" id="projectDetail" aria-hidden="true"></div>
 
   <!-- JS da aplicação -->
-  <script src="./assets/js/utils.js" defer></script>
-  <script src="./assets/js/data.js" defer></script>
-  <script src="./assets/js/templates.js" defer></script>
-  <script src="./assets/js/app.js" defer></script>
+  <script type="module" src="./assets/js/main.js"></script>
 </body>
 </html>
